rename `spi` to `stats`
