+++ setup.cfg
@@ -41,7 +41,7 @@ install_requires =
 [options.extras_require]
 test =
     pytest
-spi =
+stats =
     scipy
     numba-scipy
